requirements.txt
numpy
scipy
scikit-learn
tqdm
torchvision
soundfile